+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 3.0.0
   * Add discovery support [#64](https://github.com/singer-io/tap-harvest/pull/64)
-  # Update bookmark logic for parent-child streams
+  * Update bookmark logic for parent-child streams
   * Complete refactoring using singer_template_config
 
 ## 2.1.3